+++ tests/unit/exam/api.spec.ts
@@ -1,6 +1,5 @@
 import axios from 'axios'
 import ApiExamsService from '../../../src/services/apiExams'
-import database from '../../../data/db'
 
 jest.mock('axios')
 const mockedAxios = axios as jest.Mocked<typeof axios>
@@ -22,17 +21,4 @@ test('should return exam list', () => {
 
     return ApiExamsService.list().then( data => expect(data).toEqual(expected))
 
-})
-
-test('should save a client to the database', () => {
-    const connection = database.connection()
-    const customer = {
-        name: 'Luiz Filho',
-        cpf: '383383383-44'
-    }
-
-    const customerStored = connection.addCustomer(customer)
-    const otherConnection = database.connection()
-    const searched = otherConnection.getCustomer(customer.cpf)
-    expect(customer).toEqual(searched)
 })
